chore: check if the lock files exist before executing the post installation commands

+++ install.sh
@@ -4,13 +4,18 @@ cp ./.env.example ./.env
 
 composer install
 
-php artisan key:generate
-php artisan storage:link
-php artisan ide-helper:generate
-php artisan ide-helper:models -N
+if [ -f ./composer.lock ]; then
+    php artisan key:generate
+    php artisan storage:link
+    php artisan ide-helper:generate
+    php artisan ide-helper:models -N
+
+    touch ./database/database.sqlite
+    php artisan migrate:fresh --seed || php artisan migrate --seed
+fi
 
 pnpm install
-pnpm run build
 
-touch ./database/database.sqlite
-php artisan migrate:fresh --seed || php artisan migrate --seed
+if [ -f ./pnpm-lock.yaml ]; then
+    pnpm run build
+fi
